TCIO-144 Develop ReviewEntity and repo. Added ReviewEntity class and updated figure and user entity classes

TCIO-145  Developed review service layer (created mapper, DTO's, service). Also added interaction between review and other entities (figure, user), so changed figure's (and user's) associated files (such as mapper, DTO's, service).

TCIO-146 Developed review controller layer logic. Also updated Reviews Service and Security logic in controllers

TCIO-135 Developed reviews processing. Also overrode equals and hashCode in Figure, User, Review classes

+++ src/main/java/com/teamChallenge/controller/CategoryController.java
@@ -5,6 +5,8 @@ import com.teamChallenge.dto.response.CategoryResponseDto;
 import com.teamChallenge.entity.figure.sections.category.CategoryService;
 import com.teamChallenge.exception.CustomErrorResponse;
 import com.teamChallenge.exception.LogEnum;
+import com.teamChallenge.exception.exceptions.generalExceptions.UnauthorizedAccessException;
+import com.teamChallenge.security.AuthUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -18,7 +20,13 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
+
+/**
+ * Need update. Service use
+ */
+
 
 @RestController
 @RequestMapping("/api/categories")
@@ -29,6 +37,9 @@ public class CategoryController {
     private static final String URI_CATEGORY_WITH_ID = "/{id}";
 
     private final CategoryService categoryService;
+    private final AuthUtil authUtil;
+
+    private final String secReq = "BearerAuth";
 
     @GetMapping
     @Operation(description = "get all categories")
@@ -59,7 +70,6 @@ public class CategoryController {
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    @SecurityRequirement(name = "BearerAuth")
     @Operation(description = "create a category")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "201", description = "Created the category",
@@ -71,14 +81,16 @@ public class CategoryController {
                             schema = @Schema(implementation = CustomErrorResponse.class))}
             ),
     })
-    public CategoryResponseDto create(@RequestBody CategoryRequestDto categoryRequestDto) {
+    @SecurityRequirement(name = secReq)
+    public CategoryResponseDto create(@RequestBody CategoryRequestDto categoryRequestDto, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         CategoryResponseDto categoryResponseDto = categoryService.createCategory(categoryRequestDto);
         log.info("{}: Category (id: {}) has been added", LogEnum.CONTROLLER, categoryResponseDto.id());
         return categoryResponseDto;
     }
 
     @PutMapping(URI_CATEGORY_WITH_ID)
-    @SecurityRequirement(name = "BearerAuth")
     @Operation(description = "update a category by id")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Updated the category",
@@ -94,14 +106,16 @@ public class CategoryController {
                             schema = @Schema(implementation = CustomErrorResponse.class))}
             )
     })
-    public CategoryResponseDto update(@PathVariable String id, @RequestBody CategoryRequestDto categoryRequestDto) {
+    @SecurityRequirement(name = secReq)
+    public CategoryResponseDto update(@PathVariable String id, @RequestBody CategoryRequestDto categoryRequestDto, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         CategoryResponseDto categoryResponseDto = categoryService.updateCategory(id, categoryRequestDto);
         log.info("{}: Category (id: {}) has been updated", LogEnum.CONTROLLER, id);
         return categoryResponseDto;
     }
 
     @DeleteMapping(URI_CATEGORY_WITH_ID)
-    @SecurityRequirement(name = "BearerAuth")
     @Operation(description = "delete a category by id")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Deleted the category",
@@ -113,7 +127,10 @@ public class CategoryController {
                             schema = @Schema(implementation = CustomErrorResponse.class))}
             )
     })
-    public ResponseEntity<?> delete(@PathVariable String id) {
+    @SecurityRequirement(name = secReq)
+    public ResponseEntity<?> delete(@PathVariable String id, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         categoryService.deleteCategory(id);
         log.info("{}: Category (id: {}) has been deleted", LogEnum.CONTROLLER, id);
         return ResponseEntity.ok().build();
